+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 1.1.0 (Unreleased)
 - New command: `/leave`, to make the bot leave a discord server.
 - New command: `/versions`, shows the current version and the version of dependencies.
+- Watched List now uses an embed, meaning the length is 3 times longer.
 - Minor improvements to documentation site.
 - Improvements to server-log to make them have an uniform style.
 - Updated TibiaWiki database, fixed bug with potions price due to NPC Minzy